change all (int) and (double) to float

+++ src/de/openhpi/squash/model/Ball.java
@@ -1,10 +1,10 @@
 package de.openhpi.squash.model;
 
 public class Ball extends Movable {
-    double width;
-    double height;
+    float width;
+    float height;
 
-    public Ball(double width, double height){
+    public Ball(float width, float height){
         this.width = width;
         this.height = height;
     }
